requirements.txt
torch>=2.1
torchvision>=0.16
open-clip-torch>=2.26
faiss-cpu>=1.8.0
numpy
pandas
pillow
pyyaml
fastapi
uvicorn
pydantic
tqdm
